requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
torch==2.2.1
torchvision==0.17.1
tensorflow==2.15.0
pillow==10.2.0
matplotlib==3.8.3
numpy==1.26.4
pandas==2.2.1
